[MOD]module change_category: another design see http://doc.tikiwiki.org/tiki-index.php?page=Module+change_category
git-svn-id: https://svn.code.sf.net/p/tikiwiki/code/branches/1.10/tiki@10642 b456876b-0849-0410-b77d-98878d47e9d5

+++ modules/mod-change_category.php
@@ -1,8 +1,8 @@
 <?php
-// $Header: /cvsroot/tikiwiki/tiki/modules/mod-change_category.php,v 1.6.2.1 2007-12-03 15:57:51 sylvieg Exp $
+// $Header: /cvsroot/tikiwiki/tiki/modules/mod-change_category.php,v 1.6.2.2 2007-12-04 20:29:47 sylvieg Exp $
 
 //this script may only be included - so its better to die if called directly.
-// param: id, shy, notop
+// param: id, shy, notop, detail, categorize
 if (strpos($_SERVER["SCRIPT_NAME"],basename(__FILE__)) !== false) {
   header("location: index.php");
   exit;
@@ -42,17 +42,25 @@ if ($prefs['feature_categories'] == 'y' && (isset($_REQUEST['page']) || isset($_
   for ($i=0;$i<$num;$i++) {
     $categsid[] = $categs[$i]['categId'];
   }
-  
+
+  if (isset($_REQUEST['remove']) && in_array($_REQUEST['remove'], $categsid)) {
+	  $catObjectId = $categlib->is_categorized($cat_type, $_REQUEST['page']);
+	  $query = "delete from `tiki_category_objects` where `catObjectId`=? and `categId`=?";
+	  $result = $tikilib->query($query,array((int) $catObjectId, (int)$_REQUEST['remove']));
+  }
+
   if (isset($_REQUEST["modcatid"]) and $_REQUEST["modcatid"] == $id) {
-    $cats = $categlib->get_object_categories($cat_type, $cat_objid);
-    $catObjectId = $categlib->is_categorized($cat_type, $_REQUEST['page']);
-    $cs ="";
-    foreach ($cats as $cs) {
-      if (in_array($cs,$categsid)) {
-		  $query = "delete from `tiki_category_objects` where `catObjectId`=? and `categId`=?";
-		  $result = $tikilib->query($query,array((int) $catObjectId, (int) $cs));
-      }
-    }
+	$cs ="";
+	if (!isset($module_params['detail']) || $module_params['detail'] != 'y') {
+		$cats = $categlib->get_object_categories($cat_type, $cat_objid);
+		$catObjectId = $categlib->is_categorized($cat_type, $_REQUEST['page']);
+		foreach ($cats as $cs) {
+			if (in_array($cs,$categsid)) {
+				$query = "delete from `tiki_category_objects` where `catObjectId`=? and `categId`=?";
+				$result = $tikilib->query($query,array((int) $catObjectId, (int) $cs));
+			}
+		}
+	}
     if ($_REQUEST['modcatchange'] != 0) {
       $categlib->categorize_page($_REQUEST['page'], $_REQUEST['modcatchange']);
       $logslib->add_log('step',"changed ".$_REQUEST['page']." from $cs to ".$_REQUEST['modcatchange']);
@@ -82,7 +90,7 @@ if ($prefs['feature_categories'] == 'y' && (isset($_REQUEST['page']) || isset($_
     $smarty->assign('modname',$changecateg.$id);
   else
     $smarty->assign('modname',"change_category");
-  $smarty->assign('modcattitle',sprintf(tra('move %s in %s'),$_REQUEST['page'],$cat_parent));
+  $smarty->assign('modcattitle',sprintf(tra('Categorize %s in %s'),$_REQUEST['page'],$cat_parent));
   $smarty->assign('modcatlist',$modcatlist);
   $smarty->assign('modcatid',$id);
 }
